requirements.txt
streamlit>=1.28.0
GitPython>=3.1.40
gitignore-parser>=0.1.9